add HigherThanVenus function
